refactor(domain): split naming-rule template parser into a submodule
